Added coveo integration and search switching
