+++ package.json
@@ -13,6 +13,7 @@
     "algoliasearch": "^4.3.0",
     "app-localizer": "^1.2.2",
     "constate": "^3.1.0",
+    "coveo-search-ui": "^2.10081.2",
     "css-reset-and-normalize": "^2.1.0",
     "eslint-plugin-jsx-a11y": "^6.3.1",
     "formik": "^2.1.4",
