+++ src/AlgoliaSearch.scss
