+++ src/AlgoliaSearchBar.scss
